LayoutTests/editing/pasteboard/paste-image-as-blob-url-expected.txt

Tests that pasted images use blob URL, not webkit fake URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pastedImages.length is 1
PASS url.protocol is "blob:"
PASS successfullyParsed is true

TEST COMPLETE
